Deliver custom parallel version of cpufreqctl.
This will provide a new parallel version of `cpufreqctl` named
`cpufreqctl.auto-cpufreq` that is to be used by `auto-cpufreq`. This
prevents the system-wide version of this file from being clobbered,
which affects the integrity of what the package manager originally
installed. Applications depending on `cpufreqctl` will still get the
version that they expect.