Rework read_stats output formatting
Makes the output of the --stats argument a lot more readable.

* Formatted the output of each core properly into a table

* Colored the headers of each table column in green for easier visual parsing

* Colored the temperature reading in red for any core that is above 85 degrees so it's easy to quickly tell if something is going wrong. Might want to adjust the number it is pretty arbitrary.

This is kind of jank but only because it was already jank to begin with. The reading and writing of stats as a whole should probably be refactored to use a more convenient interface but it works fine as it is now so I'm not sure if it's actually necessary.